BRIDGE-1916: automate policies for s3 buckets
Ideally we want to manage existing BridgePF buckets with CF however AWS
does not support this use case[1]. An alternative is to let CF create
new buckets for BridgePF however that would require migrating data
from old buckets to new buckets which would be pretty painful at
this point.

Another alternative is to not manage the bucket itself
but only manage the policies for BridgePF buckets. This change adds a
user, a group, and a policy.  It applies the policy allowing the group
to access BridgePF buckets.  The compromise with this implementation
is that BridgePF S3 buckets cannot be fully automated. Users will need
to create the required buckets before running our CF template to create
the rest of the the BridgePF infrastructure.

Additional steps are required to complete this change:

1. This change will create a new bridgepf service user (with access tokens).
The new tokens are made available as EB env vars so that it is available
to the bridgepf app.
2. Bridgepf should be refactored to use this user's token to access all
buckets instead of having to get multiple tokens to access each bucket.
3. Old users for each bucket should be removed from AWS IAM. This
can be done manually in AWS console.

[1] https://forums.aws.amazon.com/thread.jspa?messageID=290625