Setup policies for BridgePF service user
The Bridge app uses the AWS heroku account to access AWS resources.
The problem is that the heroku account has too much access.  We want
to enforce a least privilege model of access.  This is ongoing work
to setup a user with a reduced set of privileges for BridgePF.

The BridgePF service user group uses many AWS services.  We need to add
additional policies to allow access to those services.

Policies added:
 SQS - full access, except for creating queues
 DDB - full access
 RDS - full access
 SNS - full access
 SES - full access
 EC  - full access, except for deletion