Switch BridgePF to use the new service user
The Bridge app uses the AWS heroku account to access AWS resources.
The problem is that the heroku account has too much access.  We want
to enforce a least privilege model of access.

Switch BridgePF to use the CF managed BridgePF service user account
to access AWS resources.  This user has less privileges than heroku
and is setup with access policies specifically for BridgePF which is
more secure.

Additional benefits:

* The management of the group, users and policies are fully automated and
changes to them are reviewable in git.
* Makes CF scripts simpler because we no longer need to pass in keys to
the template.  The access keys are completely managed in CF.
* Makes AWS IAM easier to manage because BridgePF only needs one service
account now.
* Potentially makes BridgePF simpler since it only needs one user account
to access only the AWS resources it needs.  It's easier to manage one
account than five accounts.

Optional todo:
* Refactor BridgePF to only load and use one key instaed of five keys.